// src/app/shared/models/product.ts
export class Product {
  $key: string;
  productId: string;
  productName: string;
  productCategory: string;
  productPrice: number;
  productDescription: string;
  productImageUrl: string;
  productAdded: number;
  productPtype: string;
  productQuatity: number;
  ratings: number;
  favourite: boolean;
  productSeller: string;
}